Update: pagina1 page, update: twitter page, update: main
Update pagina1 page adding the navigation in the twitter icon, it will navigate to TwitterPage.
Update twitter page turning from stateless widget to statefull widget, it with the purpose of use setState with the new property called activar the setState will be used into the onPressed method of the floating action button and the property animate of the animation of the twitter icon will be managed by the property activar, when the user pressed the button the property will change its value and the animation will start, also adding an appbar in the page to avoid the user of get back to the previous page.
Update main changing the home from the Twitter Page to Pagina1Page.